requirements.txt
google-cloud-pubsub
faker
python-dotenv